# Speed up Python apply_mask 20x by using int.from_bytes/to_bytes