--- a/src/websockets/utils.py
+++ b/src/websockets/utils.py
@@ -45,6 +45,6 @@ def apply_mask(data: bytes, mask: bytes) -> bytes:
         raise ValueError("mask must contain 4 bytes")
 
     data_int = int.from_bytes(data, sys.byteorder)
-    mask_repeated = mask * (len(data) // 4) + mask[:len(data) % 4]
+    mask_repeated = mask * (len(data) // 4) + mask[: len(data) % 4]
     mask_int = int.from_bytes(mask_repeated, sys.byteorder)
     return (data_int ^ mask_int).to_bytes(len(data), sys.byteorder)