--- a/src/websockets/utils.py
+++ b/src/websockets/utils.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import base64
 import hashlib
-import itertools
 import secrets
 import sys