@aaugustin 2021-08-19
Reading the blog post again, I considered option 4 back then (after seeing it in another library) but decided to build a C extension instead.