@benhoyt 2021-08-19
Fair push-back. And impressive work on `speedups.c` with the fancy SSE2 code!

However, I'm working on a use case where we can't use C extensions, and probably need to vendor the library as well. We could vendor it and patch it, but that gets messy. We've actually ended up going with [websocket-client](https://github.com/websocket-client/websocket-client) because we only need the client side, it's about 2/3 the size, and without C extensions does the masking 20x faster.

In terms of simplicity, a one-liner is definitely slightly nicer, but 4 straight-forward lines is not exactly terribly complex. I didn't love the `translate` implementation partly because it's non-obvious with those `[::4]` stride and `translate` operations; I find this "int xor" approach straight-forward by comparison. So to me this seems like an obvious win for those (admittedly few) people like us who can't use C extensions, and no loss for those that can.