@aaugustin 2021-08-19
OK, I think you made the right choice, I'll merge your version (with cosmetic changes).