@benhoyt 2021-08-19
Great, thank you! We haven't started implementing the Python side of this yet, so we may well change our minds on websocket-client yet. :-)